Replace parousia disc with mono for vol/pos.
Audio in Minecraft must be mono in order for it to be positional in the world. Examples of this can be seen in how vanilla Minecarts have two separate audio files, one that is mono for a cart that you're *not* riding (and can hear moving past you), and one that is stereo for a cart you *are* riding (and thus sounds like it's coming from all around you). 

Other examples include how (I believe) all (but I *know* some) music discs are mono, including Pigstep which famously has a stereo variety you can listen to *out* of the game, but in game it is intentionally mono *in* game to provide positional audio features.

Playing audio intended to be positional in stereo results in typically loud volumes as well. Replacing `parousia.ogg` with a mono variety fixes both the volume and the disc's lack of positional features.